Move Protobuf tests to separate file
